Extract to src/components/managerlib/finderutils.ts
